// src/store/user/user.types.ts
export enum USER_ACTIONS_TYPES {
  SET_CURRENT_USER = 'user/SET_CUREENT_USER',
  CHECK_USER_SESSION = 'user/CHECK_USER_SESSION',
  GOOGLE_SIGN_IN_START = 'user/GOOGLE_SIGN_IN_START',
  EMAIL_SIGN_IN_START = 'user/EMAIL_SIGN_IN_START',
  SING_IN_SUCCESS = 'user/SING_IN_SUCCESS',
  SING_IN_FAILD = 'user/SING_IN_FAILD',
  SIGN_UP_START = 'user/SIGN_UP_START',
  SIGN_UP_SUCCESS = 'user/SIGN_UP_SUCCESS',
  SIGN_UP_FAILD = 'user/SIGN_UP_FAILD',
  SIGN_OUT_START = 'user/SIGN_OUT_START',
  SIGN_OUT_SUCCESS = 'user/SIGN_OUT_SUCCESS',
  SIGN_OUT_FAILD = 'user/SIGN_OUT_FAILD',
}